Abstracted out email invitation from /invite endpoint to middleware.
